% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glia.R
\name{glia_init}
\alias{glia_init}
\title{Initialize the Glia client}
\usage{
glia_init(api_url = NULL, app_name = NULL, app_version = NULL, tags = list())
}
\arguments{
\item{api_url}{The URL of the Glia API backend. Defaults to the `GLIA_API_URL`
environment variable or "http://host:8000/ingest".}

\item{app_name}{A global name for the application or project being tracked.}

\item{app_version}{A global version for the application.}

\item{tags}{A named list of global tags to be included with every tracked event.}
}
\description{
Configures the package-level client that will be used for all tracking.
This should be called once when your application or script starts.
}
\examples{
\dontrun{
glia_init(
  api_url = "http://my-glia-instance:8000",
  app_name = "DataPipeline",
  app_version = "v2.1.0",
  tags = list(team = "data-science", priority = "high")
)
}
}
