% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glia.R
\name{glia_track}
\alias{glia_track}
\title{Track an R expression}
\usage{
glia_track(
  expr,
  name = NULL,
  version = NULL,
  tags = list(),
  description = NULL
)
}
\arguments{
\item{expr}{The R expression to execute and track.}

\item{name}{A specific name for this tracked block. Overrides the global `app_name`
and the script's filename.}

\item{version}{A specific version for this tracked block. Overrides `app_version`.}

\item{tags}{A named list of tags to add to this specific tracked event. These are
merged with any global tags set by `glia_init()`.}

\item{description}{A string providing more context for what this block does.}
}
\value{
The result of the evaluated expression.
}
\description{
Executes an R expression and records system metrics (wall time, CPU usage,
memory), sending them to the Glia backend. It uses `on.exit` to ensure
that metrics are sent even if the expression fails.
}
\examples{
\dontrun{
glia_init()

# Track a simple block of code
glia_track({
  print("Doing some work...")
  Sys.sleep(1)
  a <- 1 + 1
  print("Work done.")
}, name = "data-processing", tags = list(step = "ingestion"))

# It captures the return value
result <- glia_track({
  x <- 5 * 10
  x
})
print(result) # 50
}
}
