% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glia.R
\name{glia_wrap}
\alias{glia_wrap}
\title{Wrap a function for tracking}
\usage{
glia_wrap(f, name = NULL, version = NULL, tags = list(), description = NULL)
}
\arguments{
\item{f}{The function to wrap.}

\item{name}{A specific name for this tracked function. Defaults to the name
of the function variable.}

\item{version}{A specific version for this tracked block. Overrides `app_version`.}

\item{tags}{A named list of tags to add to this specific tracked event.}

\item{description}{A string providing more context for what this function does.}
}
\value{
A new function that automatically tracks its execution.
}
\description{
Creates a new version of a function that, when called, will be automatically
tracked by `glia_track()`.
}
\examples{
\dontrun{
glia_init()

my_heavy_computation <- function(n) {
  Sys.sleep(0.5)
  return(sum(1:n))
}

# Create a tracked version of the function
tracked_computation <- glia_wrap(my_heavy_computation, tags = list(type = "summation"))

# Calling the new function executes it and sends the metrics
result <- tracked_computation(1000)
}
}
