% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/extendr.R
\docType{package}
\name{gliar-package}
\alias{gliar-package}
\title{gliar: R Client for Glia Observability Suite}
\description{
Lightweight R client for pushing system telemetry to the Glia backend.

Lightweight R client for pushing system telemetry to the Glia backend.
}
\author{
\strong{Maintainer}: Luis Felipe Montemayor \email{lfelipemonte@gmail.com}

}
\keyword{internal}
